util: move union, difference and same functions
These functions, even though are only being used in issue_edit, are obvious
candidates for an "util" file, even more considering they were constructed in a
really generic manner.
