+++ cmd/issue_edit.go
@@ -214,59 +214,6 @@ func editText(title string, body string) (string, error) {
 	return b.String(), nil
 }
 
-// union returns all the unique elements in a and b
-func union(a, b []string) []string {
-	mb := map[string]bool{}
-	ab := []string{}
-	for _, x := range b {
-		mb[x] = true
-		// add all of b's elements to ab
-		ab = append(ab, x)
-	}
-	for _, x := range a {
-		if _, ok := mb[x]; !ok {
-			// if a's elements aren't in b, add them to ab
-			// if they are, we don't need to add them
-			ab = append(ab, x)
-		}
-	}
-	return ab
-}
-
-// difference returns the elements in a that aren't in b
-func difference(a, b []string) []string {
-	mb := map[string]bool{}
-	for _, x := range b {
-		mb[x] = true
-	}
-	ab := []string{}
-	for _, x := range a {
-		if _, ok := mb[x]; !ok {
-			ab = append(ab, x)
-		}
-	}
-	return ab
-}
-
-// same returns true if a and b contain the same strings (regardless of order)
-func same(a, b []string) bool {
-	if len(a) != len(b) {
-		return false
-	}
-
-	mb := map[string]bool{}
-	for _, x := range b {
-		mb[x] = true
-	}
-
-	for _, x := range a {
-		if _, ok := mb[x]; !ok {
-			return false
-		}
-	}
-	return true
-}
-
 func init() {
 	issueEditCmd.Flags().StringArrayP("message", "m", []string{}, "use the given <msg>; multiple -m are concatenated as separate paragraphs")
 	issueEditCmd.Flags().StringSliceP("label", "l", []string{}, "add the given label(s) to the issue")
