+++ cmd/issue_edit_test.go
@@ -5,7 +5,6 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -124,38 +123,4 @@ func Test_issueEditAssignees(t *testing.T) {
 
 	// the output should NOT show the assignee
 	require.NotContains(t, issueShowOuput, "Assignees: lab-testing")
-}
-
-func Test_issueEditSame(t *testing.T) {
-	t.Parallel()
-	assert.True(t, same([]string{}, []string{}))
-	assert.True(t, same([]string{"a"}, []string{"a"}))
-	assert.True(t, same([]string{"a", "b"}, []string{"a", "b"}))
-	assert.True(t, same([]string{"a", "b"}, []string{"b", "a"}))
-	assert.True(t, same([]string{"b", "a"}, []string{"a", "b"}))
-
-	assert.False(t, same([]string{"a"}, []string{}))
-	assert.False(t, same([]string{"a"}, []string{"c"}))
-	assert.False(t, same([]string{}, []string{"c"}))
-	assert.False(t, same([]string{"a", "b"}, []string{"a", "c"}))
-	assert.False(t, same([]string{"a", "b"}, []string{"a"}))
-	assert.False(t, same([]string{"a", "b"}, []string{"c"}))
-}
-
-func Test_issueEditUnion(t *testing.T) {
-	t.Parallel()
-	s := union([]string{"a", "b"}, []string{"c"})
-	assert.Equal(t, 3, len(s))
-	assert.True(t, same(s, []string{"a", "b", "c"}))
-}
-
-func Test_issueEditDifference(t *testing.T) {
-	t.Parallel()
-	s := difference([]string{"a", "b"}, []string{"c"})
-	assert.Equal(t, 2, len(s))
-	assert.True(t, same(s, []string{"a", "b"}))
-
-	s = difference([]string{"a", "b"}, []string{"a", "c"})
-	assert.Equal(t, 1, len(s))
-	assert.True(t, same(s, []string{"b"}))
 }
